menuModal: fix editSong link. SongEditView: fix textginput size issue
